Create example_run_CESM.sh
Original Author: Dunyu Liu
Last modified: 03/24/2021 by Dunyu Liu

This document shows a bash script from creating a CESM case to submitting the job to HPC clusters to run it. The example is for a model with the compset of B1850 and grid resolutions of f19_g17. The account charged against is A-ig2. Directories shall be changed accordingly and explanations are given following each directory. For Lonestar5, you can copy the following script into a file with a name like example_run_CESM.sh. To run the script with tcsh, type in a terminal ‘tcsh example_run_CESM.sh’. An example ‘example_run_CESM.sh’ can be found in the shared directory /work/07931/dunyuliu/group_share/ on Lonestar5 for users with permission. For permission and access to the shared directory, please contact Dr. Geeta Persad via geeta.persad@jsg.utexas.edu.