Create example_resubmit_CESM.sh
Original Author: Dunyu Liu
Last modified: 03/24/2021 by Dunyu Liu

This document shows a shell script to resubmit a job for an existing CESM case. The example is for a model with the compset of B1850 and grid resolutions of f19_g17. For Lonestar5, you can copy the following script into a file with a name like example_resubmit_CESM.sh. To run the script with tcsh, type in a terminal ‘tcsh example_resubmit_CESM.sh’. For permission and access to the shared directory, please contact Dr. Geeta Persad via geeta.persad@jsg.utexas.edu.